PC-985 Workaround for buggy ansible SSH module
Depsite no passphase being set, the module is complaining
that the key is invalid or has a passphrase. Manually verified
the key is valid. In other areas of source code for the module
it assumes no passphase is the same failure as not having a
passphrase set. Assuming there was some oversight in handling
intentionally having keys without passphrases.